+++ modules/farm_rothamsted_quick/src/Plugin/QuickForm/QuickOperation.php
@@ -39,7 +39,7 @@ class QuickOperation extends QuickExperimentFormBase {
   /**
    * {@inheritdoc}
    */
-  protected $machineryGroupNames = ['Cultivation Equipment'];
+  protected $machineryEquipmentTypes = ['Cultivation Equipment'];
 
   /**
    * {@inheritdoc}